Solved publishing issue added validation for Save in other config for folder pat

+++ DominatorUIUtility/ViewModel/OtherConfigurations/SoftwareSettingsViewModel.cs
@@ -6,6 +6,7 @@ using MahApps.Metro.Controls.Dialogs;
 using Prism.Commands;
 using System;
 using System.Diagnostics;
+using System.IO;
 using System.Windows;
 
 namespace DominatorUIUtility.ViewModel.OtherConfigurations
@@ -32,18 +33,30 @@ namespace DominatorUIUtility.ViewModel.OtherConfigurations
 
         private void Save()
         {
-            if (_softwareSettings.Save())
-            {
-                var result = Dialog.ShowCustomDialog("Success",
-                    "Software Settings sucessfully saved.To apply this setting you need to restart.\nDo you want to Restart?", "Restart now", "Restart later");
-                if (result == MessageDialogResult.Affirmative)
+            if (SoftwareSettingsModel.IsDefaultExportPathSelected && !string.IsNullOrEmpty(SoftwareSettingsModel.ExportPath))
+                if (Directory.Exists(SoftwareSettingsModel.ExportPath))
                 {
-                    Application.Current.Shutdown();
-                    Process.Start(Application.ResourceAssembly.Location);
-                    Process.GetCurrentProcess().Kill();
-                    Environment.Exit(0);
+                    if (_softwareSettings.Save())
+                    {
+                        var result = Dialog.ShowCustomDialog("Success",
+                            "Software Settings sucessfully saved.To apply this setting you need to restart.\nDo you want to Restart?", "Restart now", "Restart later");
+                        if (result == MessageDialogResult.Affirmative)
+                        {
+                            Application.Current.Shutdown();
+                            Process.Start(Application.ResourceAssembly.Location);
+                            Process.GetCurrentProcess().Kill();
+                            Environment.Exit(0);
+                        }
+                    }
                 }
-            }
+                else
+                {
+                    SoftwareSettingsModel.ExportPath = string.Empty;
+                    
+                    Dialog.ShowDialog("Error", "Please enter valid folder Path.");
+                }
+
+
         }
     }
 }
